+++ release-notes.md
@@ -6,6 +6,8 @@
 - Centralize driver error handling
 - Fixes #51, #48 - Halin errors on dbms.security.auth_enabled=false
 - Fixes #49 - Halin errors in some cases when a non-admin user is used for login
+- Fixes data display issues in the store file sizes component to make them human readable
+- Fixes bug where address of machine involved in role changes did not update properly
 
 ## 0.6.0